+++ atlasdb-client/src/main/java/com/palantir/atlasdb/transaction/impl/CachingTransaction.java
@@ -15,17 +15,21 @@
  */
 package com.palantir.atlasdb.transaction.impl;
 
+import java.util.Arrays;
 import java.util.Map;
 import java.util.Set;
 import java.util.SortedMap;
-import java.util.concurrent.ConcurrentMap;
+import java.util.concurrent.ExecutionException;
+import java.util.stream.Collectors;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.google.common.cache.Cache;
 import com.google.common.cache.CacheBuilder;
 import com.google.common.cache.CacheLoader;
 import com.google.common.cache.LoadingCache;
+import com.google.common.cache.Weigher;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.ImmutableSortedMap;
@@ -33,6 +37,7 @@ import com.google.common.collect.Iterables;
 import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
 import com.google.common.primitives.UnsignedBytes;
+import com.google.common.util.concurrent.UncheckedExecutionException;
 import com.palantir.atlasdb.encoding.PtBytes;
 import com.palantir.atlasdb.keyvalue.api.Cell;
 import com.palantir.atlasdb.keyvalue.api.ColumnSelection;
@@ -40,24 +45,36 @@ import com.palantir.atlasdb.keyvalue.api.RowResult;
 import com.palantir.atlasdb.keyvalue.api.TableReference;
 import com.palantir.atlasdb.keyvalue.impl.Cells;
 import com.palantir.atlasdb.transaction.api.Transaction;
+import com.palantir.common.base.Throwables;
 
 public class CachingTransaction extends ForwardingTransaction {
-
     private static final Logger log = LoggerFactory.getLogger(CachingTransaction.class);
 
-    final Transaction delegate;
+    private static final int DEFAULT_MAX_CACHE_SIZE = 100;
+    private static final int DEFAULT_MAX_CACHE_WEIGHT_PER_TABLE = 10_000_000;
 
-    private final LoadingCache<String, ConcurrentMap<Cell, byte[]>> columnTableCache = CacheBuilder.newBuilder()
-            .softValues()
-            .build(new CacheLoader<String, ConcurrentMap<Cell, byte[]>>() {
-        @Override
-        public ConcurrentMap<Cell, byte[]> load(String key) throws Exception {
-            return Maps.newConcurrentMap();
-        }
-    });
+    final Transaction delegate;
+    private final int maxCacheWeightPerTable;
+
+    // The keys here are Strings for historic reasons, as TableReferences treat empty namespace and the met namespace
+    // the same.
+    private final Cache<String, LoadingCache<Cell, byte[]>> columnTableCache;
+
+    public CachingTransaction(Transaction delegate, int maxCacheSize, int maxCacheWeightPerTable) {
+        this.delegate = delegate;
+        this.maxCacheWeightPerTable = maxCacheWeightPerTable;
+        this.columnTableCache = CacheBuilder.newBuilder()
+                .maximumSize(maxCacheSize)
+                .removalListener(notification -> {
+                    if (log.isDebugEnabled() && notification.wasEvicted()) {
+                        log.debug("Caching transaction cache for table {} was evicted.", notification.getKey());
+                    }
+                })
+                .build();
+    }
 
     public CachingTransaction(Transaction delegate) {
-        this.delegate = delegate;
+        this(delegate, DEFAULT_MAX_CACHE_SIZE, DEFAULT_MAX_CACHE_WEIGHT_PER_TABLE);
     }
 
     @Override
@@ -67,13 +84,13 @@ public class CachingTransaction extends ForwardingTransaction {
 
     @Override
     public SortedMap<byte[], RowResult<byte[]>> getRows(TableReference tableRef, Iterable<byte[]> rows,
-                                                        ColumnSelection columnSelection) {
+            ColumnSelection columnSelection) {
         if (Iterables.isEmpty(rows)) {
             log.info("Attempted getRows on '{}' table and {} with empty rows argument", tableRef, columnSelection);
             return AbstractTransaction.EMPTY_SORTED_ROWS;
         }
 
-        ConcurrentMap<Cell, byte[]> colCache = getColCacheForTable(tableRef);
+        LoadingCache<Cell, byte[]> colCache = getColCacheForTable(tableRef);
         if (columnSelection.allColumnsSelected()) {
             SortedMap<byte[], RowResult<byte[]>> loaded = super.getRows(tableRef, rows, columnSelection);
             cacheLoadedRows(colCache, loaded.values());
@@ -88,7 +105,7 @@ public class CachingTransaction extends ForwardingTransaction {
                 boolean nonEmpty = false;
                 boolean shouldLoad = false;
                 for (byte[] col : columnSelection.getSelectedColumns()) {
-                    byte[] val = colCache.get(Cell.create(row, col));
+                    byte[] val = colCache.getIfPresent(Cell.create(row, col));
                     if (val == null) {
                         shouldLoad = true;
                         break;
@@ -110,8 +127,8 @@ public class CachingTransaction extends ForwardingTransaction {
         }
     }
 
-    private void cacheLoadedRows(ConcurrentMap<Cell, byte[]> colCache,
-                                 Iterable<RowResult<byte[]>> rowView) {
+    private void cacheLoadedRows(Cache<Cell, byte[]> colCache,
+            Iterable<RowResult<byte[]>> rowView) {
         for (RowResult<byte[]> loadedRow : rowView) {
             for (Map.Entry<Cell, byte[]> e : loadedRow.getCells()) {
                 cacheLoadedColumns(colCache, ImmutableSet.of(e.getKey()), ImmutableMap.of(e.getKey(), e.getValue()));
@@ -119,18 +136,21 @@ public class CachingTransaction extends ForwardingTransaction {
         }
     }
 
-    private void cacheLoadedColumns(ConcurrentMap<Cell, byte[]> colCache,
-                                    Set<Cell> toLoad,
-                                    Map<Cell, byte[]> toCache) {
+    private void cacheLoadedColumns(Cache<Cell, byte[]> colCache,
+            Set<Cell> toLoad,
+            Map<Cell, byte[]> toCache) {
         for (Cell key : toLoad) {
             byte[] value = toCache.get(key);
             if (value == null) {
                 value = PtBytes.EMPTY_BYTE_ARRAY;
             }
-            colCache.putIfAbsent(key, value);
+            colCache.put(key, value);
         }
     }
 
+    // Cache semantics: Zero byte values represent a deleted cell, which the delegate will return to us as null, but
+    // we will put in the cache as a zero-length array to differentiate "we saw a deleted cell" from "this cell isn't
+    // in the cache".
     @Override
     public Map<Cell, byte[]> get(TableReference tableRef, Set<Cell> cells) {
         if (cells.isEmpty()) {
@@ -138,29 +158,18 @@ public class CachingTransaction extends ForwardingTransaction {
             return ImmutableMap.of();
         }
 
-        ConcurrentMap<Cell, byte[]> cache = getColCacheForTable(tableRef);
-        Set<Cell> toLoad = Sets.newHashSet();
-        Map<Cell, byte[]> cacheHit = Maps.newHashMapWithExpectedSize(cells.size());
-        for (Cell cell : cells) {
-            byte[] val = cache.get(cell);
-            if (val != null) {
-                if (val.length > 0) {
-                    cacheHit.put(cell, val);
-                }
-            } else {
-                toLoad.add(cell);
-            }
+        try {
+            return getColCacheForTable(tableRef).getAll(cells).entrySet()
+                    .stream()
+                    .filter(entry -> entry.getValue().length > 0)
+                    .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
+        } catch (ExecutionException | UncheckedExecutionException e) {
+            throw Throwables.throwUncheckedException(e.getCause());
         }
-
-        final Map<Cell, byte[]> loaded = super.get(tableRef, toLoad);
-
-        cacheLoadedColumns(cache, toLoad, loaded);
-        cacheHit.putAll(loaded);
-        return cacheHit;
     }
 
     @Override
-    final public void delete(TableReference tableRef, Set<Cell> cells) {
+    public final void delete(TableReference tableRef, Set<Cell> cells) {
         super.delete(tableRef, cells);
         addToColCache(tableRef, Cells.constantValueMap(cells, PtBytes.EMPTY_BYTE_ARRAY));
     }
@@ -172,18 +181,65 @@ public class CachingTransaction extends ForwardingTransaction {
     }
 
     private void addToColCache(TableReference tableRef, Map<Cell, byte[]> values) {
-        Map<Cell, byte[]> colCache = getColCacheForTable(tableRef);
-        for (Map.Entry<Cell, byte[]> e : values.entrySet()) {
-            byte[] value = e.getValue();
-            if (value == null) {
-                value = PtBytes.EMPTY_BYTE_ARRAY;
-            }
-            colCache.put(e.getKey(), value);
+        LoadingCache<Cell, byte[]> colCache = getColCacheForTable(tableRef);
+        colCache.putAll(setMissingKeysToEmpty(values, values.keySet()));
+    }
+
+    private LoadingCache<Cell, byte[]> getColCacheForTable(TableReference tableRef) {
+        String tableName = tableRef.getQualifiedName();
+        try {
+            return columnTableCache.get(tableName, () -> createNewCacheForTable(tableRef));
+        } catch (ExecutionException e) {
+            throw Throwables.throwUncheckedException(e.getCause());
         }
     }
 
-    private ConcurrentMap<Cell, byte[]> getColCacheForTable(TableReference tableRef) {
-        String tableName = tableRef.getQualifiedName();
-        return columnTableCache.getUnchecked(tableName);
+    private LoadingCache<Cell, byte[]> createNewCacheForTable(TableReference tableRef) {
+        return CacheBuilder.newBuilder()
+                .softValues()
+                .maximumWeight(maxCacheWeightPerTable)
+                .weigher((Weigher<Cell, byte[]>) (key, value) -> value.length)
+                .removalListener(notification -> {
+                    if (log.isDebugEnabled() && notification.wasEvicted()) {
+                        String rowName = notification.getKey() == null ? "null cell" :
+                                Arrays.toString(notification.getKey().getRowName());
+                        String colName = notification.getKey() == null ? "null cell" :
+                                Arrays.toString(notification.getKey().getColumnName());
+                        log.debug("Removing entry from caching transaction cache: Table {}, Row {}, Column {},"
+                                        + " Eviction reason {}", tableRef.getQualifiedName(), rowName, colName,
+                                notification.getCause().toString());
+                    }
+                })
+                .build(getCacheLoader(tableRef));
+    }
+
+    private CacheLoader<Cell, byte[]> getCacheLoader(TableReference tableRef) {
+        return new CacheLoader<Cell, byte[]>() {
+            @Override
+            public byte[] load(Cell key) throws Exception {
+                return setMissingKeysToEmpty(CachingTransaction.super.get(tableRef, ImmutableSet.of(key)),
+                        ImmutableSet.of(key))
+                        .get(key);
+            }
+
+            @Override
+            public Map<Cell, byte[]> loadAll(Iterable<? extends Cell> keys) throws Exception {
+                Set<Cell> keySet = ImmutableSet.copyOf(keys);
+                return setMissingKeysToEmpty(CachingTransaction.super.get(tableRef, ImmutableSet.copyOf(keys)), keySet);
+            }
+        };
+    }
+
+    private Map<Cell, byte[]> setMissingKeysToEmpty(Map<Cell, byte[]> returnedMap, Set<Cell> expectedKeys) {
+        Map<Cell, byte[]> res = Maps.newConcurrentMap();
+        for (Cell key : expectedKeys) {
+            byte[] value = returnedMap.get(key);
+            if (value == null) {
+                res.put(key, PtBytes.EMPTY_BYTE_ARRAY);
+            } else {
+                res.put(key, value);
+            }
+        }
+        return res;
     }
 }